requirements.txt
Django>=6.0
python-decouple>=3.8
django-allauth>=65.0.0
requests>=2.32.0
PyJWT>=2.10.0
cryptography>=45.0.0
